# fix: Convert folding recursive verifier ops to batch mul