Updates folding recursive verifier to use batch_mul for optimal goblin ec op efficiency. This reduces a single recursive verification from 1144 ECC ops to 264. (Note: 264 = 6*44 where 6 is the number of ecc op gate rows needed for two scalar muls (one for each instance) plus an "equals" op and 44 is the number of witnesses plus precomputed polys, not including shifts)

Closes https://github.com/AztecProtocol/barretenberg/issues/849

New benchmark result:
```
-----------------------------------------------------------------
Benchmark                       Time             CPU   Iterations
-----------------------------------------------------------------
IvcBench/Full/6             54156 ms        51691 ms            1
```
Old benchmark result:
```
-----------------------------------------------------------------
Benchmark                       Time             CPU   Iterations
-----------------------------------------------------------------
IvcBench/Full/6             66891 ms        63569 ms            1
```